Move LayerFilters up one level
